split methods to modules
